Sua diem HoSo.java

+++ HoSo.java
@@ -1,6 +1,6 @@
  class HoSo {
  	String Ten;
- 	Integer Diem;
+ 	double  Diem;
